refactor: move cache and encryption to core, providers to llm module
